chore: Move some files around
- We added a `utils.ts` file, but that might get merged into the `data.ts` file.
