--- mobile/src/modules/customization/theme/helpers/zod.ts
@@ -1,49 +1,0 @@
-import { z } from "zod/mini";
-
-import { ZSchema } from "~/modules/form/utils";
-import { normalizeHexColor } from "./color";
-import type { ColorRole } from "../core/constants";
-import { ColorRoleOptions } from "../core/constants";
-
-export const HexColorSchema = z.pipe(
-  ZSchema.NonEmptyString,
-  z.transform((str, ctx) => {
-    const normalized = normalizeHexColor(str);
-    if (normalized) return normalized;
-
-    ctx.issues.push({
-      code: "invalid_value",
-      input: str,
-      values: [str],
-      message: "Expected a valid hex color.",
-    });
-    return z.NEVER;
-  }),
-);
-
-const ColorRolesSchema = Object.fromEntries(
-  ColorRoleOptions.map((role) => [role, HexColorSchema]),
-) as Record<ColorRole, typeof HexColorSchema>;
-
-//#region Form Schema
-export const ThemeEntrySchema = z.object({
-  // Additional context:
-  _importGen: z.nullable(z.number()),
-  // Actual form fields:
-  name: ZSchema.NonEmptyString,
-  scheme: z.enum(["light", "dark"]),
-  ...ColorRolesSchema,
-});
-
-export type ThemeEntry = z.infer<typeof ThemeEntrySchema>;
-//#endregion
-
-//#region Backup Schema
-export const ThemeExportSchema = z.object({
-  name: ZSchema.NonEmptyString,
-  scheme: z.enum(["light", "dark"]),
-  colors: z.object(ColorRolesSchema),
-});
-
-export type ThemeExport = z.infer<typeof ThemeExportSchema>;
-//#endregion
