+++ mobile/src/modules/customization/theme/screens/View.tsx
@@ -19,13 +19,13 @@ import { FlatList } from "~/components/Base/List";
 import { Pressable } from "~/components/Base/Pressable";
 import { FilledIconButton, IconButton } from "~/components/Form/Button/Icon";
 import { StyledText } from "~/components/Typography/StyledText";
-import { exportTheme } from "../helpers/backup";
 import type { CustomTheme, HexColor } from "../core/constants";
 import {
   DefaultThemeOptions,
   Themes as DefaultThemes,
   SystemTheme,
 } from "../core/constants";
+import { exportTheme } from "../core/data";
 import { useCustomThemes } from "../queries";
 import { formatAsCustomTheme, isDefaultTheme } from "../utils";
